For some reason, g++ wants a cast...
git-svn-id: svn+ssh://svn.code.sf.net/p/yabause/code/trunk@3151 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/qt/ui/UIDebugSH2.cpp
@@ -180,7 +180,7 @@ void UIDebugSH2::updateTrackInfLoop()
          twTrackInfLoop->setItem(i, 0, newItem);
 
          newItem = new QTableWidgetItem();
-         newItem->setData(Qt::DisplayRole, match[i].count);
+         newItem->setData(Qt::DisplayRole, (qulonglong) match[i].count);
          twTrackInfLoop->setItem(i, 1, newItem);
       }
       twTrackInfLoop->setSortingEnabled(true);
